Trial session smoke tests - add retries
