+++ cypress-smoketests.json
@@ -12,5 +12,6 @@
   "defaultCommandTimeout": 20000,
   "requestTimeout": 20000,
   "viewportWidth": 1200,
-  "viewportHeight": 900
+  "viewportHeight": 900,
+  "retries": 3
 }
